Refactor App.tsx by responsibility
App.tsx had grown to ~1000 lines mixing server I/O, the live-preview sync
state machine, pure helpers, and the full render. Split it into focused
modules:

- lib/previewServer.ts — preview-server protocol (network calls + types)
- lib/clipboard.ts / fileKey.ts / stableStringify.ts / settings.ts — small
  single-responsibility modules (replacing a grab-bag helpers file)
- hooks/usePreviewSync.ts — the publish/share state machine (token+revision
  bookkeeping, cold-start reconcile, conflict handling, share actions)
- components/PresetsTab.tsx (+ module css) — the presets-browsing tab UI

App.tsx is now the composition root (~320 lines): app state, bridge wiring,
and tab routing. Behavior is unchanged; SyncStatus is re-exported from App so
existing imports keep working.
