+++ figma/src/ui/components/PresetsTab.module.css
@@ -1,0 +1,62 @@
+/* Presets tab — the scrollable browse view: search + grouped controls card,
+   the results toolbar (count / sound / layout), the preset list, and the
+   floating scroll-to-top button. */
+
+.preset-scroll { flex: 1; }
+
+/* Grouped controls card — Filters / Custom presets / Phone share one bordered
+   panel below the search box, replacing the old loose accordion rows. */
+.controls-section {
+  padding: 8px;
+  border-bottom: 1px solid var(--border);
+}
+.controls-search { position: relative; }
+.controls-search-clear { position: absolute; right: 4px; top: 4px; margin: 0; }
+.controls-card {
+  margin-top: 8px;
+  border: 1px solid var(--color-blue-20);
+  border-radius: var(--radius);
+  overflow: hidden;
+  background: var(--surface);
+}
+/* Neutralise the standalone .accordion divider/padding so rows sit flush in the
+   card, separated by hairlines instead. */
+.controls-card > :global(.acc-row) { border-top: 1px solid var(--color-blue-10); padding-top: 0; }
+.controls-card > :global(.acc-row):first-child { border-top: none; }
+
+/* Results toolbar — count + sound toggle + layout switch. */
+.list-toolbar { padding: 4px 8px; gap: 6px; margin-top: 8px; }
+.list-toolbar-meta { font-size: var(--fs-xs); }
+.list-toolbar-toggle {
+  gap: 6px;
+  margin-left: 8px;
+  font-size: var(--fs-xs);
+  cursor: pointer;
+  user-select: none;
+}
+.preset-list { padding-top: 8px; }
+.preset-list-empty { padding: 16px; }
+
+/* Floating scroll-to-top, mirrors the docs preset-list button. */
+button.scroll-top {
+  position: fixed;
+  right: 16px;
+  bottom: 16px;
+  z-index: 1000;
+  width: 44px;
+  height: 44px;
+  padding: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 2px solid var(--color-blue-50);
+  border-radius: var(--radius);
+  background: var(--surface);
+  color: var(--color-blue-50);
+  box-shadow: var(--shadow-card);
+}
+button.scroll-top:hover {
+  box-shadow: -4px 4px 0 var(--color-blue-50);
+  transform: translate(1px, -1px);
+}
+button.scroll-top:active { box-shadow: none; transform: translate(-3px, 3px); }
