+++ figma/src/ui/lib/clipboard.ts
@@ -1,0 +1,18 @@
+// Clipboard inside the Figma plugin iframe: navigator.clipboard is often blocked,
+// so fall back to a temporary textarea + execCommand('copy').
+export function copyToClipboard(text: string): boolean {
+  try {
+    const ta = document.createElement('textarea');
+    ta.value = text;
+    ta.style.position = 'fixed';
+    ta.style.opacity = '0';
+    document.body.appendChild(ta);
+    ta.focus();
+    ta.select();
+    const ok = document.execCommand('copy');
+    document.body.removeChild(ta);
+    return ok;
+  } catch {
+    return false;
+  }
+}
