+++ figma/src/ui/lib/fileKey.ts
@@ -1,0 +1,5 @@
+// Accept either a raw file key or a full Figma URL and return the key.
+export function extractFileKey(input: string): string {
+  const m = input.match(/figma\.com\/(?:file|design|proto)\/([A-Za-z0-9]+)/);
+  return (m ? m[1] : input).trim();
+}
