+++ figma/src/ui/lib/settings.ts
@@ -1,0 +1,8 @@
+import type { Settings } from '../../shared/types';
+
+export const DEFAULT_SETTINGS: Settings = {
+  soundInEdit: true,
+  compactLayout: false,
+  fileKeyOverride: '',
+  previewBaseUrlOverride: ''
+};
